+++ src/roq/samples/example-8/strategy.cpp
@@ -106,7 +106,7 @@ void Strategy::create_order(MessageInfo const &message_info, Layer const &layer)
       .order_type = OrderType::LIMIT,
       .time_in_force = TimeInForce::GTC,
       .execution_instructions = {},
-      .order_template = {},
+      .request_template = {},
       .quantity = quantity_,
       .price = price,
       .stop_price = NaN,