+++ src/roq/samples/event-log/processor.cpp
@@ -124,6 +124,10 @@ void Processor::operator()(Event<CustomMetricsUpdate> const &event) {
   log::info("{}"sv, event);
 }
 
+void Processor::operator()(Event<CustomMatrixUpdate> const &event) {
+  log::info("{}"sv, event);
+}
+
 void Processor::operator()(Event<ParametersUpdate> const &event) {
   log::info("{}"sv, event);
 }