+++ src/roq/samples/vwap/processor.hpp
@@ -42,6 +42,7 @@ struct Processor final : public client::EventLogReader::Handler {
   void operator()(Event<PositionUpdate> const &) override {}
   void operator()(Event<FundsUpdate> const &) override {}
   void operator()(Event<CustomMetricsUpdate> const &) override {}
+  void operator()(Event<CustomMatrixUpdate> const &) override {}
   void operator()(Event<ParametersUpdate> const &) override {}
   void operator()(Event<PortfolioUpdate> const &) override {}