+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changes
+
+* All order requests must support `request_template` (#329)
+
 ## 0.9.3 &ndash; 2023-03-20
 
 ## 0.9.2 &ndash; 2023-02-22