+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes will be documented in this file.
 
 ## Head
 
+## 0.9.4 &ndash; 2023-05-04
+
 ### Changes
 
 * Install Mambaforge (#351)